# docs/swagger.yaml
basePath: /
definitions:
...
  Blogs.CreateUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - password
    - role
    - username
    type: object
...
  Blogs.SignUpUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - password
    - username
    type: object
...
  Blogs.UpdateUserRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  handler.TokenRoleResponse:
    properties:
      role:
        type: string
      token:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
host: localhost:8086
info:
  contact: {}
  description: Blogs Server in Go using Gin framework
  title: Blogs Server API
  version: "1.0"
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      operationId: get-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - ApiKeyAuth: []
      summary: Get All Users
      tags:
      - Administration
    post:
      consumes:
      - application/json
      description: Create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Blogs.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - ApiKeyAuth: []
      summary: Create User
      tags:
      - Administration
  /admin/users/{users_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing User by ID
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: users_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete an existing User
      tags:
      - Administration
    get:
      consumes:
      - application/json
      description: Get a single User by providing its ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: users_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a User by ID
      tags:
      - Administration
    put:
      consumes:
      - application/json
      description: Update an existing User data with new data
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: users_id
        required: true
        type: string
      - description: new user info
        in: body
        name: new_input
        required: true
        schema:
          $ref: '#/definitions/Blogs.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/http.Response'
      security:
      - ApiKeyAuth: []
      summary: Update an existing User data
      tags:
      - Administration
  /auth/admin/sign-in:
    post:
      consumes:
      - application/json
      description: Login to admin account
      operationId: login-admin-account
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign In Admin
      tags:
      - Administration
  /auth/author/sign-in:
    post:
      consumes:
      - application/json
      description: Login to author account
      operationId: login-author-account
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign In Author
      tags:
      - Authorization
  /auth/author/sign-up:
    post:
      consumes:
      - application/json
      description: Create account for author
      operationId: create-account-author
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Blogs.SignUpUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign Up Author
      tags:
      - Authorization
  /auth/user/sign-in:
    post:
      consumes:
      - application/json
      description: Login to user account
      operationId: login-user-account
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign In User
      tags:
      - Authorization
  /auth/user/sign-up:
    post:
      consumes:
      - application/json
      description: Create account for user
      operationId: create-account-user
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Blogs.SignUpUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign Up User
      tags:
      - Authorization
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"